audiosync-yt/requirements.txt
flask==2.3.3
werkzeug==2.3.7
yt-dlp==2023.7.6
flask-cors==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
requests